:recycle: api: moving calculateCertificationAssessmentScore next to scoring V2
